- updated to SF rev. 303
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@840 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Horde3D/Source/Horde3DUtils/main.cpp
@@ -903,7 +903,7 @@ DLLEXP bool h3dutScreenshot( const char *filename )
 	delete[] pixels;
 	h3dutFreeMem( &image );
 
-	return bytesWritten == width * height * 3 + 18;
+	return bytesWritten == (size_t) width * height * 3 + 18;
 }
